requirements.txt
google-genai>=1.47.0
Pillow>=11.0.0